feat: suporte para interiro de 64 bits

+++ main.cpp
@@ -1,20 +1,21 @@
 #include <iostream>
+#include <cstdint>
 
 using namespace std;
 
-int fibonacciRecursivo(int n);
+ uint64_t fibonacciRecursivo(int n);
 
 int main() {
     int n;
     cout << "digite o valor do termo da funcao fibonacci:";
     cin >> n;
-    int f = fibonacciRecursivo(n);
+    uint64_t f = fibonacciRecursivo(n);
     cout << "\nfibonacci(" << n << ")=" << f << endl;
     return 0;
 }
 
 // função recursiva para calcular o n-ésimo termo da sequência de fibonacci, complexidade exponêncial 2^n
-int fibonacciRecursivo(int n){
+uint64_t fibonacciRecursivo(int n){
     if(n<=1){
         return n;
     }
